refactor(support): split persistence -- move recovery store + SessionLock to DanTermSupport
Phase 4 of the pure-core/portable-support refactor. Extracts the recovery-path
helpers (recoveryDirectoryURL/light/enriched/sessionLockURL), the session-lock IO
(write/read/deleteSessionLockFile), and the SessionLock type out of DanTermCore into
a new DanTermSupport/RecoveryStore.swift, and drops CoreEnv.recoveryDir. The lock
functions lose their `env: CoreEnv` parameter and instead take defaulted
recoveryDir/now/pid seams -- a CoreEnv dependency would be the forbidden support->core
edge. Production call sites stay zero-arg, so app code is byte-for-byte unchanged and
resolves the funcs same-module via the app/DanTermSupport symlink. The 5 path/lock tests
split out of CheckpointTests into RecoveryStoreTests (round-trip reworked off makeTestEnv
to the explicit seams, pinning a fixed pid). Pure codec/merge/validation stays in core.
Zero public annotations.

No app changes, no manifest changes, no .gitignore change, no test-ui.sh path edit.
CLI/IPC surface unchanged, so no SKILL.md edit. just test + just build + just test-ui
green (core 1054/36 -> 1049/36; support 11/2 -> 16/3; protocol 121 unchanged; UI 12/12).
